Move standalone tests and jobrunner into combined package
