--- executor/mock/mock.go
@@ -1,1 +1,0 @@
-package mock
